# src/main/resources/static/openapi.yaml
openapi: 3.1.0
info:
  title: codeAnalyzer API
  version: 1.0.0
  description: >
    Clone repositories, analyze Java code (parse → embed → store in Chroma),
    query the indexed codebase, and interact via natural language chat agent.
servers:
  - url: http://localhost:8080
paths:
...
  /api/chat/message:
    post:
      summary: Send a message to the chat agent
      description: >
        Natural language interface for code analysis. The agent can analyze repositories,
        find endpoints, search code, and answer questions about the codebase.
      tags:
        - Chat Agent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [message]
              properties:
                message:
                  type: string
                  description: Natural language message or question
                  example: "Analyze https://github.com/spring-projects/spring-boot"
                sessionId:
                  type: string
                  description: Session ID for conversation context (auto-generated if not provided)
                  example: "chat_session_12345"
      responses:
        "200":
          description: Chat response
          content:
            application/json:
              schema:
                type: object
                properties:
                  response:
                    type: string
                    description: Conversational response from the agent
                    example: "I'll analyze the Spring Boot repository for you..."
                  suggestions:
                    type: array
                    items:
                      type: string
                    description: Suggested follow-up questions or actions
                    example: ["What REST endpoints does it have?", "Show me the main components"]
                  success:
                    type: boolean
                    description: Whether the operation was successful
                    example: true
                  sessionId:
                    type: string
                    description: Session ID for this conversation
                    example: "chat_session_12345"
                  metadata:
                    type: object
                    description: Additional metadata about the operation
                    properties:
                      actionExecuted:
                        type: string
                        description: The action that was performed
                        example: "ANALYZE_REPO"
                      actionSuccess:
                        type: boolean
                        description: Whether the action succeeded
                        example: true
                      dataAvailable:
                        type: boolean
                        description: Whether additional data is available
                        example: true
        "500":
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  response:
                    type: string
                    example: "I'm having trouble processing your request. Please try again."
                  success:
                    type: boolean
                    example: false
                  sessionId:
                    type: string
  /api/chat/session/{sessionId}/context:
    get:
      summary: Get conversation context for a session
      description: Retrieve the conversation history and context for a specific chat session
      tags:
        - Chat Agent
      parameters:
        - in: path
          name: sessionId
          required: true
          schema:
            type: string
          description: The session ID to get context for
          example: "chat_session_12345"
      responses:
        "200":
          description: Session context
          content:
            application/json:
              schema:
                type: object
                properties:
                  sessionId:
                    type: string
                    example: "chat_session_12345"
                  currentRepository:
                    type: string
                    nullable: true
                    description: Currently analyzed repository URL
                    example: "https://github.com/spring-projects/spring-boot"
                  currentRepositoryId:
                    type: string
                    nullable: true
                    description: Internal repository identifier
                    example: "spring-projects_spring-boot"
                  recentExchanges:
                    type: array
                    description: Recent conversation exchanges (max 5)
                    items:
                      type: object
                      properties:
                        userMessage:
                          type: string
                          example: "What endpoints does this have?"
                        assistantResponse:
                          type: string
                          example: "I found 15 REST endpoints..."
                        timestamp:
                          type: string
                          format: date-time
                          example: "2024-01-15T10:30:00Z"
                  lastActivity:
                    type: string
                    format: date-time
                    description: Last activity timestamp
                    example: "2024-01-15T10:35:00Z"
                  sessionData:
                    type: object
                    description: Additional session-specific data
                    additionalProperties: true
        "404":
          description: Session not found
        "500":
          description: Server error
...